Adds basic Feature Label display to the viewer
Colouring along makes distinguishing clusters difficult.
This adds (togglable) display labels to the Viewer. An
API was added to fetch the required data (name, colour,
location) from the backend. As part of this typing was
also improved.

This also fixes a bug in GProfiler, adds documentation,
tests, and some typing on the server.

Closes #78