Get loom metadata on feature update
This causes the frontend to receive possible the latest annotations and metadata (such as votes or new annotations) when switching features. This means that when multiple people are working on the same file such as during a collaborative annotation session, all users see updates in real time without a reload required.