Initial template for user, project, and dataset APIs
This introduces no-op functionality into the existing SCope
application. No "wiring up" has been done. _All_ data is stored in the
database. It is treated as an alternative filesystem so existing Loom
handling code will still work minus the file handling complexity. The
added advantage is that multiple instances of SCope will run on a
single machine.

Users are identified by an integer that the client must store in a
Cookie. Projects have 2 alternative identifiers: an integer and a
uuid-v4. Both of these are generated on the server. Each project
stores its size and creation date so that a cron job can clean up as
necessary. This means that clients no longer need to poll the server.