Remove pyscenic dependency and increase Python version
PyScenic was used for "gene set enrichment" that was previously disabled. So
this change removes the gene set enrichment code and also the PyScenic
dependency. Because PyScenic was pulling in a lot of dependencies this
also allows us to update to Python 3.9.

Unfortunately we're still stuck on an old version of h5py because of loompy so
there are no h5py wheels for Python 3.9.